refactor: relocating the iothub clients
